Armstrong Number
Def: In number theory, a narcissistic number or “Armstrong number” is a number where in a given number base b is a number that is the sum of its own digits each raised to the power of the number of digits. 
This definition is way more complicated than it actually is. 
E.g. :  153=1^3+5^3+3^3